change datatype according to sql

+++ projects/project_0/Project0.StoreApplication.Domain/Models/Product.cs
@@ -2,10 +2,10 @@ namespace Project0.StoreApplication.Domain.Models
 {
   public class Product
   {
-    public int ProductId { get; set; }
+    public byte ProductId { get; set; }
 
     public string Name { get; set; }
-    public double Price { get; set; }
+    public decimal Price { get; set; }
 
     public override string ToString()
     {
